[FLINK-7343][kafka-tests] Fix test at-least-once test instability
Previously we could set lastSnapshotedElement to some value during checkpointing
AFTER executing shutdown while KafkaProducer snapshot of this value would fail.
This was leading to incorrectly expect this value to be present in the test kafka topic.
Fix is to remember lastSnapshotedElementBeforeShutdown - last snapshot that we exepct
to succeed without failure.